requirements.txt

# Core
pydantic>=2.5.0
pydantic-ai>=1.33
pydantic-settings>=2.1.0

# OpenAI
openai>=1.12.0

# PyTorch & Vision
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Training utilities
tqdm>=4.66.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database & Persistence
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Logging & Monitoring
rich>=13.7.0
loguru>=0.7.2

# CLI
click>=8.1.7
typer>=0.9.0

# Async
asyncio>=3.4.3
aiohttp>=3.9.0
tenacity>=8.2.3

# Utilities
psutil>=5.9.0
GPUtil>=1.4.0